requirements.txt
Django>=4.2.23,<5.0
python-docx>=1.1.0
PyPDF2>=3.0.1
requests>=2.31.0
python-dotenv>=1.0.0
celery>=5.3.4
redis>=5.0.1
Pillow>=10.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7 